+++ src/components/Account/AccountBalance.scss
@@ -1,0 +1,30 @@
+.displaybalance {
+  display: flex;
+  flex-direction: row-reverse;
+  max-width: 33rem;
+  margin-bottom: 1rem;
+  .empty{
+    width: 60%;
+  }
+  .verified {
+    width: 5%;
+    margin-left: 1rem;
+    margin-right: 1rem;
+    margin-top: 0.4em;
+  }
+  .name {
+    width: 10%;
+    margin-right: 1rem;
+    margin-top: 0.4em;
+  }
+  .balance{
+    width: 50%;
+  }
+  .remove{
+    width: 5%;
+    cursor: pointer;
+    &:hover {
+      opacity: 1;
+    }
+  }
+}
