+++ src/components/ErrorView/ErrorView.scss
