+++ src/components/Loader/loader.scss
