+++ src/styles.scss
@@ -256,53 +256,6 @@ section > .field:not(:last-child) {
   border-radius: 0.25rem;
 }
 
-.account {
-  display: flex;
-  max-width: 33rem;
-  .address {
-    width: 70%;
-    margin-right: 1rem;
-  }
-  .result{
-    width: 30%;
-    display: flex;
-    span {
-      opacity: 0.5;
-    }
-    .nonce{
-      margin-right: 1rem;
-    }
-    .icon{
-      padding-top: 1.8rem;
-    }
-  }
-}
-
-.displaybalance {
-  display: flex;
-  flex-direction: row-reverse;
-  max-width: 33rem;
-  margin-bottom: 0rem;
-  .empty{
-    width: 60%;
-  }
-  .name {
-    width: 10%;
-    margin-right: 1rem;
-    margin-top: 0.4em;
-  }
-  .balance{
-    width: 30%;
-  }
-  .remove{
-    width: 5%;
-    cursor: pointer;
-    &:hover {
-      opacity: 1;
-    }
-  }
-}
-
 .displaycontract {
     max-width: 33rem;
     opacity: 0.5;
